corrected issues with styling
corrected an issue with the footer being too close to the bottom of the viewport. Also corrected multiple issues with the carousel. Fixed the name and descriptions covering the images as well as fixed a bottom padding issue where the active carousel indicator would cut off the bottom of the description.